+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>21.6.83-pre-prod.1 (2022-05-11)</small>
+
+* Merge pull request #616 from WTW-IM/add-drawer-item-onchange ([0b20dfe](https://github.com/WTW-IM/es-components/commit/0b20dfe)), closes [#616](https://github.com/WTW-IM/es-components/issues/616)
+
+
+
+
+
 ## <small>21.6.83-pre-prod.0 (2022-05-10)</small> & <small>21.6.83 (2022-05-10)</small>
 
 ## <small>21.6.82 (2022-01-31)</small>